+++ pkg/dev_compiler/lib/src/js_ast/nodes.dart
@@ -304,6 +304,7 @@ abstract class Node {
   }
 }
 
+// TODO(jmesserly): rename to Module.
 class Program extends Node {
   /// Script tag hash-bang, e.g. `#!/usr/bin/env node`
   final String scriptTag;