+++ pkg/dev_compiler/test/codegen_expected/BenchmarkBase.js
@@ -1,6 +1,4 @@
-dart_library.library('BenchmarkBase', null, /* Imports */[
-  'dart_sdk'
-], function load__BenchmarkBase(exports, dart_sdk) {
+define(['dart_sdk'], function(dart_sdk) {
   'use strict';
   const core = dart_sdk.core;
   const dart = dart_sdk.dart;
@@ -92,5 +90,7 @@ dart_library.library('BenchmarkBase', null, /* Imports */[
     names: ['measureFor']
   });
   // Exports:
-  exports.BenchmarkBase = BenchmarkBase$;
+  return {
+    BenchmarkBase: BenchmarkBase$
+  };
 });