+++ pkg/dev_compiler/tool/build_sdk.sh
@@ -9,14 +9,14 @@ dart -c tool/patch_sdk.dart tool/input_sdk gen/patched_sdk
 echo "*** Compiling SDK to JavaScript"
 
 # TODO(jmesserly): break out dart:html & friends.
-#
-# Right now we can't summarize our SDK, so we can't treat it as a normal
-# explicit input (instead we're implicitly compiling against the user's SDK).
-#
-# Another possible approach is to hard code the dart:* library->module mapping
-# into the compiler itself, so it can emit the correct import.
-#
 dart -c tool/build_sdk.dart \
     --dart-sdk gen/patched_sdk \
-    -o lib/runtime/dart_sdk.js \
+    --modules=amd \
+    -o lib/js/amd/dart_sdk.js \
+    --modules=es6 \
+    -o lib/js/es6/dart_sdk.js \
+    --modules=common \
+    -o lib/js/common/dart_sdk.js \
+    --modules=legacy \
+    -o lib/js/legacy/dart_sdk.js \
     "$@" > tool/sdk_expected_errors.txt